backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.16.1
requests==2.32.3
python-dotenv==1.0.1
algoliasearch>=4.0,<5.0
gunicorn==22.0.0
brevo-python>=1.2.0
a2wsgi==1.10.4
uvicorn[standard]==0.30.6

